% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/metric-learning.R
\name{metric.learning}
\alias{metric.learning}
\title{Metric learning between source data and desired target}
\usage{
metric.learning(target, df)
}
\arguments{
\item{target}{A vector (or single column data frame) that is the desired targets.}

\item{df}{Numeric data frame where features are columns.}
}
\description{
This function computes the weighted metric between a source data frame and a desired target.
}
\examples{
metric.learning()
x <- read.table(textConnection('t0 t1 t2
aaa  0  1  0
bbb  1  0  1
ccc  1  1  1
ddd  1  1  0
' ), header=TRUE)
metric.learning(target = x$t0, df = x)
}
\keyword{correlation}
\keyword{distances}
\keyword{pairwise}
\keyword{row}

