% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pairwise-dist.R
\name{pairwise.dist}
\alias{pairwise.dist}
\title{Pairwise Distance Data Frame Computation}
\usage{
pairwise.dist(key, df, dist.matrix = FALSE, method = "euclidean", p = 2)
}
\arguments{
\item{key}{A unique identifier key for each row}

\item{df}{Numeric data frame (or matrix) to compute pairwise distances of rows}

\item{dist.matrix}{Defaults to FALSE. Set to TRUE if the provided data is a pre-computed distance matrix}

\item{method}{Defaults to "euclidean" same as R dist function method. Ignored if dist.matrix = TRUE}

\item{p}{Defaults to 2 same R dist function used in Minkowski distance. Ignored if dist.matrix = TRUE}
}
\description{
This function takes a data frame and computes the distance between ever pair of rows.
}
\examples{
pairwise.dist()
x <- read.table(textConnection('t0 t1 t2
aaa  0  1  0
bbb  1  0  1
ccc  1  1  1
ddd  1  1  0
' ), header=TRUE)
pairwise.dist(key = rownames(x), df = x)
}
\keyword{distances}
\keyword{pairwise}
\keyword{row}

