% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/pairwise-dist.R
\name{target.source.diff}
\alias{target.source.diff}
\title{Pairwise Distance Comparison between Target and Source data frames}
\usage{
target.source.diff(target.key, target.df, source.key, source.df,
  target.matrix = FALSE, source.matrix = FALSE,
  target.method = "euclidean", target.p = 2, source.method = "euclidean",
  source.p = 2)
}
\arguments{
\item{target.key}{A unique identifier key for each row target data frame row. Must match the row key for the source data frame}

\item{target.df}{Numeric target data frame (or matrix) to compute pairwise distances of rows}

\item{source.key}{A unique identifier key for each row source data frame row. Must match the row key for the target data frame}

\item{source.df}{Numeric source data frame (or matrix) to compute pairwise distances of rows}

\item{target.matrix}{Defaults to FALSE. Set to TRUE if the provided target data is a pre-computed distance matrix}

\item{source.matrix}{Defaults to FALSE. Set to TRUE if the provided source data is a pre-computed distance matrix}

\item{target.method}{Defaults to "euclidean" same as R dist function method. Ignored if target.matrix = TRUE}

\item{target.p}{Defaults to 2 same R dist function used in Minkowski distance. Ignored if target.matrix = TRUE}

\item{source.method}{Defaults to "euclidean" same as R dist function method. Ignored if source.matrix = TRUE}

\item{source.p}{Defaults to 2 same R dist function used in Minkowski distance. Ignored if source.matrix = TRUE}
}
\description{
This function computes the row pairwise distances for a target and a source dataframe and merges them based on a unique id row key.
}
\examples{
pairwise.dist()
x <- read.table(textConnection('t0 t1 t2
aaa  0  1  0
bbb  1  0  1
ccc  1  1  1
ddd  1  1  0
' ), header=TRUE)
target.source.diff(target.key = rownames(x), target.df = x, source.key = rownames(x), source.df = x)
}
\keyword{correlation}
\keyword{distances}
\keyword{pairwise}
\keyword{row}

